refactor constants files
